restructuring of the files, and the problem of the loop generated by the function of adding a client was solved

+++ src/person.h
@@ -4,6 +4,6 @@ using namespace std;
 
 struct Person{
     // TODO Hacer que el dni sea unico en la lista
-    long int dni;
+    long long int dni;
     string name;
 };
